refactor: 뷰에서 수행하던 녹화 기능을 RecordManager 타입으로 분리
